#33 Incorrect code generated for property Map<String, Bar>
