+++ README.md
@@ -7,7 +7,7 @@ Simply specify the object graph root class and BuilderGenerator will generate a 
 consistent pattern for each JavaBean in the entire object graph.
 
 ##Features
-aaa
+
 - Generate a complete set of builders for an entire object graph with a single line of code
 - Non intrusive and time saving 
   - no need to annotate your code
